+++ patzilla/navigator/services/ops.py
@@ -5,7 +5,7 @@ from cornice.service import Service
 from pyramid.httpexceptions import HTTPNotFound
 from pyramid.settings import asbool
 from patzilla.access.epo.ops.api import ops_published_data_crawl, inquire_images, get_ops_image, ops_family_inpadoc, get_ops_client, pdf_document_build, ops_claims, ops_document_kindcodes, ops_description, ops_published_data_search, ops_published_data_search_invalidate, ops_published_data_search_swap_family, ops_family_publication_docdb_xml
-from patzilla.navigator.services import propagate_keywords, cql_prepare_query
+from patzilla.navigator.services import propagate_keywords, cql_prepare_query, handle_generic_exception
 from patzilla.access.generic.exceptions import NoResultsException
 from patzilla.util.numbers.common import split_patent_number
 from patzilla.util.python import _exception_traceback
@@ -95,15 +95,18 @@ def ops_published_data_search_handler(request):
         else:
             result = ops_published_data_search(constituents, query, range)
         propagate_keywords(request, query_object)
+        return result
 
     except NoResultsException as ex:
         # Forward response to let the frontend recognize zero hits
         request.response.status = HTTPNotFound.code
         return ex.data
 
-    log.info('query finished')
+    except Exception as ex:
+        message = handle_generic_exception(request, ex, 'ops-search', query)
+        request.errors.add('ops-search', 'search', message)
 
-    return result
+    log.info('query finished')
 
 
 @ops_published_data_crawl_service.get(accept="application/json")