+++ CHANGES.rst
@@ -13,8 +13,9 @@ development
 - Pull application-wide upstream API authentication credentials from datasource settings
 - Trim down automatic user provisioning
 - Application configuration file cleanup
-- Namespace refactoring
+- Large namespace refactoring
 - Make tests work again
+- Properly handle and propagate cache database connection errors
 
 2017-09-08 0.135.0
 ==================